+++ apps/bsadmin/src/app/(pages)/(set)/svc-cate-list/[...slug]/page.tsx
